Added Phpmyadmin
Still working out the kinks with the google load balancer, which is acting bizzare.
I can get Apache2 and phpmyadmin to work separately.  If I combine the paths one the ingress, I can get apache2 to work on http, but not https, and when I try to navigate to /phpmyadmin, it says there's no backend